feat: Improve user and home navigation, add expense creation feedback (#dev)
The changes in this commit aim to improve user navigation and provide better
feedback mechanisms.

In the Home page, checks were added to redirect users to the 'sign-in' page if
they are not currently logged in. If the user is logged in and has a homeId,
they are redirected to their home page.

In the CreateExpense component, the 'useParams' and 'useToast' hooks were
added. 'useParams' is used to obtain the homeId from the URL, and 'useToast'
provides a toast notification to the user upon successful creation of an
expense.

Changes were also made to the 'prisma.expense.create' function in
'src/app/actions/expense/index.ts' to revalidate the home page after an
expense is created, ensuring the data displayed is up to date.

In 'src/app/actions/user/index.ts', caching was applied to the